chore(accordion): moves interface to component file
