--- packages/accordion/src/types.ts
@@ -1,17 +1,0 @@
-import type { ComponentProps, ReactNode } from 'react';
-
-export interface AccordionRootProps extends ComponentProps<'div'> {
-  children: ReactNode;
-  asChild?: boolean;
-}
-export interface AccordionItemProps {
-  children: ReactNode;
-  className?: string;
-  defaultOpen?: boolean;
-}
-
-export interface AccordionContentProps {
-  children: ReactNode;
-  className?: string;
-  asChild?: boolean;
-}
